fix(reviews): nightly cron self-heals a wrong gbpLocationName (Tamarind→Shah Alam)
The scoreboard showed Tamarind with Shah Alam's 1,486 reviews and a fake
160.6/day velocity. Snapshot history dates the poisoning precisely: Tamarind's
rows were correct (353→370 reviews, ~17/30d) until Jul 2; on Jul 3 its count
jumped to Shah Alam's 1,485 — the day the wrong locations/… id took effect.

- lib/reviews/relink.ts: relinkGbpLocations(apply) extracted from the
  /api/reviews/gbp-relink route (route now a thin wrapper).
- reviews-daily-snapshot cron: runs the repair (apply mode) before
  snapshotting, so a provable placeId mismatch self-heals the same night
  instead of silently poisoning snapshots, the review feed, auto-reply and
  the relevance audit. Best-effort — a relink failure never blocks snapshots.
- The three poisoned Tamarind snapshot rows (Jul 3–5, wrong listing's counts)
  were deleted from prod; the scoreboard falls back to the last true row and
  rebuilds from the next snapshot.
